+++ source/CPUGPUMemory.h
@@ -13,15 +13,15 @@ private:
 	void *memCPU;
 	void *memGPU;
 
-	int size;
+	size_t size;
 
 	bool is_float;
 	GPU *gpu;
 public:
-	CPUGPUMemory(GPU *_gpu, bool _is_float, int _size, float _initValues);
+	CPUGPUMemory(GPU *_gpu, bool _is_float, size_t _size, float _initValues);
 	~CPUGPUMemory();
 
-	void Resize(int newSize);
+	void Resize(size_t newSize);
 	void* GetCPUMemory();
 	void* GetGPUMemory();
 	int GetSize();
